# Conflicts:
#	flexget/plugins/filter/series.py

+++ flexget/plugins/output/download.py
@@ -1,6 +1,6 @@
 from __future__ import unicode_literals, division, absolute_import
 
-from builtins import *  # pylint: disable=unused-import, redefined-builtin
+from builtins import *  # noqa pylint: disable=unused-import, redefined-builtin
 from future.moves.urllib.parse import unquote
 
 import hashlib
@@ -481,6 +481,8 @@ class PluginDownload(object):
                         raise plugin.PluginError('Unable to write %s: %s' % (destfile, err))
                     if err.errno != errno.EPERM and err.errno != errno.EACCES:
                         raise
+                else:
+                    del(entry['file'])
 
             # store final destination as output key
             entry['location'] = destfile
@@ -501,7 +503,8 @@ class PluginDownload(object):
             if os.path.exists(entry['file']):
                 log.debug('removing temp file %s from %s', entry['file'], entry['title'])
                 os.remove(entry['file'])
-            shutil.rmtree(os.path.dirname(entry['file']))
+            if os.path.exists(os.path.dirname(entry['file'])):
+                shutil.rmtree(os.path.dirname(entry['file']))
             del (entry['file'])
 
     def cleanup_temp_files(self, task):
